+++ src/resources/index.ts
@@ -14,7 +14,6 @@ export {
   type CustomerRetrieveResponse,
   type CustomerUpdateResponse,
   type CustomerListResponse,
-  type CustomerRetrieveBillingResponse,
   type CustomerCreateParams,
   type CustomerUpdateParams,
   type CustomerListParams,