+++ tests/api-resources/customers.test.ts
@@ -7,11 +7,11 @@ const client = new Flowglad({
   baseURL: process.env['TEST_API_BASE_URL'] ?? 'http://127.0.0.1:4010',
 });
 
-describe('resource customerProfiles', () => {
+describe('resource customers', () => {
   // skipped: tests are disabled for the time being
   test.skip('create: only required params', async () => {
-    const responsePromise = client.customerProfiles.create({
-      customerProfile: { email: 'email', externalId: 'externalId' },
+    const responsePromise = client.customers.create({
+      customer: { email: 'email', externalId: 'externalId', name: 'name' },
     });
     const rawResponse = await responsePromise.asResponse();
     expect(rawResponse).toBeInstanceOf(Response);
@@ -24,22 +24,23 @@ describe('resource customerProfiles', () => {
 
   // skipped: tests are disabled for the time being
   test.skip('create: required and optional params', async () => {
-    const response = await client.customerProfiles.create({
-      customerProfile: {
+    const response = await client.customers.create({
+      customer: {
         email: 'email',
         externalId: 'externalId',
+        name: 'name',
         archived: true,
         domain: 'domain',
         iconURL: 'iconURL',
         logoURL: 'logoURL',
-        name: 'name',
+        userId: 'userId',
       },
     });
   });
 
   // skipped: tests are disabled for the time being
   test.skip('retrieve', async () => {
-    const responsePromise = client.customerProfiles.retrieve('externalId');
+    const responsePromise = client.customers.retrieve('externalId');
     const rawResponse = await responsePromise.asResponse();
     expect(rawResponse).toBeInstanceOf(Response);
     const response = await responsePromise;
@@ -51,7 +52,7 @@ describe('resource customerProfiles', () => {
 
   // skipped: tests are disabled for the time being
   test.skip('update: only required params', async () => {
-    const responsePromise = client.customerProfiles.update('externalId', { customerProfile: { id: 'id' } });
+    const responsePromise = client.customers.update('externalId', { customer: { id: 'id' } });
     const rawResponse = await responsePromise.asResponse();
     expect(rawResponse).toBeInstanceOf(Response);
     const response = await responsePromise;
@@ -63,8 +64,8 @@ describe('resource customerProfiles', () => {
 
   // skipped: tests are disabled for the time being
   test.skip('update: required and optional params', async () => {
-    const response = await client.customerProfiles.update('externalId', {
-      customerProfile: {
+    const response = await client.customers.update('externalId', {
+      customer: {
         id: 'id',
         archived: true,
         domain: 'domain',
@@ -73,13 +74,14 @@ describe('resource customerProfiles', () => {
         iconURL: 'iconURL',
         logoURL: 'logoURL',
         name: 'name',
+        userId: 'userId',
       },
     });
   });
 
   // skipped: tests are disabled for the time being
   test.skip('list', async () => {
-    const responsePromise = client.customerProfiles.list();
+    const responsePromise = client.customers.list();
     const rawResponse = await responsePromise.asResponse();
     expect(rawResponse).toBeInstanceOf(Response);
     const response = await responsePromise;
@@ -93,19 +95,7 @@ describe('resource customerProfiles', () => {
   test.skip('list: request options and params are passed correctly', async () => {
     // ensure the request options are being passed correctly by passing an invalid HTTP method in order to cause an error
     await expect(
-      client.customerProfiles.list({ cursor: 'cursor', limit: 1 }, { path: '/_stainless_unknown_path' }),
+      client.customers.list({ cursor: 'cursor', limit: 1 }, { path: '/_stainless_unknown_path' }),
     ).rejects.toThrow(Flowglad.NotFoundError);
-  });
-
-  // skipped: tests are disabled for the time being
-  test.skip('retrieveBilling', async () => {
-    const responsePromise = client.customerProfiles.retrieveBilling('externalId');
-    const rawResponse = await responsePromise.asResponse();
-    expect(rawResponse).toBeInstanceOf(Response);
-    const response = await responsePromise;
-    expect(response).not.toBeInstanceOf(Response);
-    const dataAndResponse = await responsePromise.withResponse();
-    expect(dataAndResponse.data).toBe(response);
-    expect(dataAndResponse.response).toBe(rawResponse);
   });
 });